feat: fully migrating ts into js
