+++ .eslintrc.json
@@ -4,7 +4,6 @@
     "es2021": true
   },
   "extends": ["airbnb-base", "prettier"],
-  "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": "latest",
     "sourceType": "module"
@@ -16,7 +15,7 @@
       }
     }
   },
-  "plugins": ["@typescript-eslint", "prettier"],
+  "plugins": ["prettier"],
   "rules": {
     "prettier/prettier": [
       "error",
