--- global.d.ts
@@ -1,20 +1,0 @@
-declare module '*.jpg' {
-  const value: any;
-  export = value;
-}
-
-declare module '*.png' {
-  const value: any;
-  export = value;
-}
-
-declare module '*.svg?inline' {
-  const content: any;
-  export default content;
-}
-declare module '*.svg' {
-  const content: any;
-  export default content;
-}
-
-declare module '*.css';
