+++ package.json
@@ -23,11 +23,12 @@
   },
   "homepage": "https://github.com/DesPetrikov/RSS-virtual-keyboard#readme",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
+    "@babel/core": "^7.17.10",
+    "@babel/preset-env": "^7.17.10",
     "@webpack-cli/generators": "^2.4.2",
+    "babel-loader": "^8.2.5",
     "css-loader": "^6.7.1",
-    "eslint": "^8.14.0",
+    "eslint": "^8.15.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
@@ -37,8 +38,6 @@
     "sass": "^1.50.1",
     "sass-loader": "^12.6.0",
     "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.8",
-    "typescript": "^4.6.3",
     "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.8.1"
