--- src/commonTypes/dataType.types.ts
@@ -1,6 +1,0 @@
-export type dataType = {
-  ru: { key: string; shift: string };
-  en: { key: string; shift: string };
-  code: string;
-  image?: string;
-};
