--- src/commonTypes/language.types.ts
@@ -1,1 +1,0 @@
-export type languageType = 'en' | 'ru';
