+++ src/components/TextArea/TextArea.js
@@ -1,13 +1,19 @@
 import createHtml from '../../createHtml';
 import styles from './TextArea.moodule.css';
-var TextArea = /** @class */ (function () {
-    function TextArea(parent) {
-        this.parent = parent;
-        this.createTextArea();
-    }
-    TextArea.prototype.createTextArea = function () {
-        this.container = createHtml('textarea', styles.textarea, this.parent, '', 'autofocus, true');
-    };
-    return TextArea;
-}());
-export { TextArea };
+
+export class TextArea {
+  constructor(parent) {
+    this.parent = parent;
+    this.createTextArea();
+  }
+
+  createTextArea() {
+    this.container = createHtml(
+      'textarea',
+      styles.textarea,
+      this.parent,
+      '',
+      'autofocus, true'
+    );
+  }
+}
